Alignment with the 0.13.1 release of AMB

+++ src/fuel-stop-advisor/fuel-stop-advisor.cpp
@@ -350,7 +350,7 @@ class FuelStopAdvisor
         DBus::MessageIter it;
         uint32_t odometer;
         uint16_t level;
-        uint16_t consumption;
+        uint32_t consumption;
         double time;
